+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Abstractions/Users/IUserProvisioningService.cs
@@ -1,0 +1,9 @@
+using Command = TimeSheet.Modules.EmploymentManagement.Application.Users.Create.Command;
+using Result = TimeSheet.Modules.EmploymentManagement.Application.Users.Create.Result;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Abstractions.Users;
+
+public interface IUserProvisioningService
+{
+    Task<Result> CreateAsync(Command command, CancellationToken cancellationToken);
+}