+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Users/Create/Command.cs
@@ -1,0 +1,3 @@
+namespace TimeSheet.Modules.EmploymentManagement.Application.Users.Create;
+
+public sealed record Command(Guid RoleId, Guid? EmployeeId, string? Email);