+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Domain/Auditing/AuditActionTypes.cs
@@ -11,6 +11,10 @@ public static class AuditActionTypes
     public const string AccessDenied = "AccessDenied";
     public const string AccountLockedOut = "AccountLockedOut";
     public const string PasswordChanged = "PasswordChanged";
+    public const string UserCreated = "UserCreated";
+    public const string TemporaryPasswordIssued = "TemporaryPasswordIssued";
+    public const string TemporaryPasswordResent = "TemporaryPasswordResent";
+    public const string PasswordResetRequested = "PasswordResetRequested";
     public const string UserRoleAssigned = "UserRoleAssigned";
     public const string RoleAssignmentRejected = "RoleAssignmentRejected";
     public const string EmployeeCreated = "EmployeeCreated";